[*** INITIAL COMMIT ***] Initial commit for the lts6 overlay.
Initial commit for the lts6 overlay.

The lts6 is intended to support a project to port long-term-stable (lts)
packages from Enterprise Linux sources.
The '6' is for version 6 of the Scientific Linux sources used
initially for this overlay.
Scientific Linux is based on RHEL (RedHat Enterprise Linux).

The initial commit consists of the repo_name file.