[eclass] Add a supplemental rpm eclass.
Add lts6-rpm.eclass, an eclass intended to supplement the rpm
functions in the Gentoo rpm.eclass library.

The initial commit consists of a change to the rpm_spec_epatch function
to omit EPATCH_OPTS.  This was found to interfer with the initial
sys-devel/which ebuild, and is assumed to affect other ebuilds also.