[media-libs/jpeg] Recreate the Gentoo ebuild for media-libs/jpeg/jpeg-6b-r8
*** NOTE ***
This is *NOT* a SRPM ebuild.

This ebuild recreates the last retired version of jpeg-6b,
which is the same version used for Scientific Linux 6.1.

Initial attempts at building a SRPM-based jpeg-b6 ebuild have run
into trouble (primarily sandbox violations).

I'm going to step back, and take the process in a couple of chunks.
The first step is to recreate the retired Gentoo ebuild.