[sys-libs/zlib] Add SRPM-based ebuilds for sys-libs/zlib-1.2.3
Create ebuilds SRPM-based ebuilds for sys-libs/zlib-1.2.3

The upstream SRPM is from Scientific Linux 6.1:
zlib-1.2.3-26.el6.src.rpm

Two ebuilds are included in this commit.
The r26 ebuild uses a build scheme more closely associated with the
original Gentoo ebuild for version 1.2.3.

The r27 ebuild takes advantage of the SRPM patches to allow autotools
to be used to build the package.