[x11-wm/icewm] SRPM-based ebuild for x11-wm/icewm/icewm-1.2.37
Add a SRPM-based ebuild for x11-wm/icewm/icewm-1.2.37 based on the
current Gentoo ebuild for the same release.
The SRPM is from Scientific Linux 6.1:
icewm-1.2.37-1.2.src.rpm

This SRPM seems to be developed and maintained by Scientific Linux,
and is not found in the RHEL 6.1 sources.

TO DO: The portion of the ebuld that inserts the additional icons
provided by the SRPM is failing silently.

It's not a big deal since the default icons will work fine.