[dev-libs/openssl] Fix lib installation.
The last version of the ebuild had a silient error that omitted the
installation of the openssl and crypt so library files.

This was tracked down to the SRPM openssl-1.0.0-beta4-redhat.patch.
Examination of this patch showed it shouldn't be necessary since
it primarily set RPM variables that are not used.