[eclass/lts6-rpm] Fix silent error for wrong spec file name.
The processing of the SRPM spec file would fail when the spec file
name was supplied incorrectly.  However, the ebuild would continue
to build the package without applying any SRPM patches.

The eclass function will check for the existence of the supplied
file name for the spec file, and fail if not found.