[net-misc/openssh] Add a SRPM-based ebuild for net-misc/openssh-5.3_p1
Add a SRPM-based ebuild for net-misc/openssh-5.3_p1.

This ebuild is barely working, and selinux is sure to be broken.

At least the following To Do's are in order:

To Do: The fips patches require a package called 'fipscheck-devel',
which is available as a SRPM but currently doesn't exist anywhere
in Gentoo-land.

To Do:  After incorporating the 'fipscheck-devel' package, demonstrate
the 'elfips' USE flag works.

To Do:  selinux audit capabilities require a package called 'audit-libs',
also not available in Gentoo-land.  Audit may be mandatory for selinux.
So test to see if it should be part of selinux requirements, or get
it's own USE flag.

To Do:  Test selinux in general, and on the openssh package specifically.