[virtual/libusb] Add compatibility ebuild for virtual/libusb.
This is a compatibility ebuild to resolve a conflict with the
current version of virtual/libusb in portage, which restricts
dev-libs/libusb:1 to >=dev-libs/libusb-1.0.9_rc:1

The only package that really needs version 1.0.9 is usbredir
according to the Changelog.

To Do:
This solution needs further thought.

Actually, I'm not understanding why there is a virtual package
for libusb if the virtual package restricts the dependency to only
dev-libs/libusb.

And it seems broken to have your virtual include a constraint that
applies to all reverse dependencies when only one package requires
the dependency.