[app-text/ghostscript-gpl] Initial baseline commit for ghostscript-gpl.
This is an initial baseline commit of the retired Gentoo-based ebuild
for app-text/ghostscript-gpl-8.70

This commit is provided for the purpose of tracking changes made
in developing a SRPM-based ebuild.

It is planned to remove the support for djvu in the SRPM-based ebuild.
Since the Long Term Stable overlay will not be able to provide ongoing
support/bug-tracking for the djvu option, it will be left for users
interested in that option to build a current version of
app-text/ghostscript-gpl that includes a recent version of the djvu
option.