[app-text/ghostscript-gpl] Add EL SRPM-based ebuild for ghostscript-gpl
Add Enterprise Linux SRPM-based ebuild for
app-text/ghostscript-gpl-8.70-r11.

The SRPM was sourced from Scientific Linux 6.2
ghostscript-8.70-11.el6_2.6.src.rpm

Support for djvu was omitted due to the high probability of missed
maintenance fixes and bug fixes in the context of the Long Term
Stable overlay.

Users desiring this feature are encouraged to use the version on the
main Gentoo portage tree.