[x11-libs/libX11] Add EL SRPM-based ebuild for x11-libs/libX11.
Add Enterprise Linux SRPM-based Ebuild for x11-libs/libX11-1.3-r2.

The SRPM was sourced from Scientific Linux 6.2.
libX11-1.3-2.el6.src.rpm

Notes:
I ran into a dependency conflict when I merged in the line
"XORG_EAUTORECONF=yes" from the x11-libs/libX11-1.4.5 ebuild.

It should be safe to disable this in the 1.3 version, but I
want to document the issue somewhat since it may come up again.

It appears the xorg-2 is trying to set a higher version dependency
>=x11-misc/util-macros-1.17 when XORG_EAUTORECONF==yes and when
the arch is "x86-interix ppc-aix x86-winnt"

It seems like this should not have affected this ebuild on x86,
but it was causing a dependency conflict on x11-misc/util-macros.