[sys-apps/which] Updates to ebuild.
Several general updates:
(1) This was an early ebuild in developing the Long Term Stable
overlay, and the package version reflects early attempts to
identify version reconciliation between the Gentoo tree, and the
EL sources.

The version assigned to this package was not initially consistent.

(2) Update to EAPI 4 and the most recent practices for applying
SRPM patches.

(3) Upstream has adopted the interesting practice of changing the
SRPM files slightly, without changing revision.

The content appears the same, but as far as I can tell, they've
adopted a different hash scheme for signing their packages.