[media-libs/tiff] General Updates to ebuild.
The early ebuilds committed to the lts6 overlay are being
reviewed and updated as necessary to reflect the conventions
that have been accumulated as this project has progressed.

(1) Moved patching to the correct ebuild phase.

(2) Bumped EAPI from 3 to 4.

(3) Updated the jpeg dependency to provide for the package to
depend on the media-libs/jpeg.  More details on this are provided
in the subsequent virtual/jpeg commit.