[virtual/jpeg] Temporary Commit for LTS6 Compatability.
This virtual is a temporary commit to work around incompatabilities
introduced by changes to the upstream Gentoo ebuild for this virtual.

In order to address packages that rely on the advanced features
in media-libs/libjpeg-turbo, Gentoo is working with the main
Gentoo portage tree version of virtual/jpeg to accommodate this
issue.

Also, Gentoo is effectively moving on from the version of media-libs/jpeg
that is provided in the LTS6 overlay.

This commit allows the LTS6 overlay to override the main Gentoo
portage tree version of virtual/jpeg.

This is a TEMPORARY solution to allow for a long term solution to
be implemented in the LTS6 overlay.

I've settled on the following general constraints for a solution
on the LTS6 overlay:

LTS6 packages should be permitted to utilize the version of
media-libs/jpeg-6 from this overlay.

However, users that begin to bring in packages from the main Gentoo
portage tree that depend on jpeg should be forced to adopt Gentoo's
solution as of that time.

Two possible solutions are:
(1) Change each LTS6 ebuild that depends on jpeg to allow either:
|| ( media-libs/jpeg:0 virtual/jpeg )

(2) Add a new virtual/jpeg-lts6 that depends on either:
|| ( media-libs/jpeg:0 virtual/jpeg )
and update the LTS6 ebuilds to depend on that virtual.

I'm leaning towards solution (2) as it provides a single place
to react to any future changes in the upstream jpeg virtual that
conflicts with a current solution.