[virtual/jpeg-lts6] Compatibility Virtual.
This virtual is provided for compatibility with the version of
media-libs/jpeg in the LTS6 overlay.

Ebuilds in the LTS6 overlay will depend on this virtual, or the
virtual in the main Gentoo portage tree.

If users decide to bring in packages that depend on jpeg from the
main Gentoo tree, they will either have to upgrade according to
Gentoo's dependencies, or roll their own solution.