[media-libs/imlib] General Updates to Ebuild.
The early ebuilds committed to the lts6 overlay are being
reviewed and updated as necessary to reflect the conventions
that have been accumulated as this project has progressed.

(1) Updated to explicit identification of SRPM patches.

(2) Clarified that the EL local-hack-gmodule patch is not
being handled.

(3) Updated the jpeg dependency to provide for the LTS6 overlay
version of media-libs/jpeg using the lts6 virtual.