[x11-libs/gtk+:2] General Updates to Ebuild.
The early ebuilds committed to the lts6 overlay are being reviewed
and updated as necessary to reflect the conventions that have been
accumulated as this project has progressed.

(1) Updated to explicit identification of SRPM patches.

(2) Update the jpeg dependency to provide for compatability with
the older version of jpeg standard in EL 6.x (media-libs/jpeg-6b).