[dev-libs/openssl] Update to partially address build issues.
This ebuild is encountering build issues on the amd64 arch.
It looks like the EL Patches may impose the expectation of being
built with the EL Fips modifications enabled, and with +binutils
to disable patent encumbered algorithms.

I've also added back a EL patch that was omitted to provide for
hardening of the build.

This required an additional patch to revert the library naming
injected in the EL patch.