[virtual/pkgconfig] LTS6 Compatability Ebuild.
The version of virtual/pkgconfig in the Gentoo tree has set a
dependency that would disallow the version of dev-util/pkgconfig
in the lts6 overlay.  This is due to a small subset of ebuilds
that cannot work with the older pkgconfig.

So there is little choice except to fork a version of
virtual/pkgconfig for the lts6 overlay that will allow the version
know to work with the packages in this overlay.