[media-libs/jpeg] Ebuild clean-ups.
The early ebuilds committed to the lts6 overlay are being reviewed
and updated as necessary to reflect the conventions that have been
accumulated as this project has progressed.

TO DO:
This ebuild contains a over-sized patch that belongs as a separate
source file.

Currently, that resource is not in place.