[app-arch/xz-utils] Add EL SRPM-based ebuild for app-arch/xz-utils.
Add Enterprise Linux SRPM-based Ebuild for
app-arch/xz-utils-4.999.9_beta-r11.

The SRPM was sourced from Scientific Linux 6.2.
xz-4.999.9-0.3.beta.20091007git.el6.src.rpm

Notes:
I had to implement a work-around for unpacking the source package.

The SRPM source tar archive for xz-utils is itself in a tar.xz
archive.  This presents the dilema of needing to assume you have
the ability to unpack an xz archive when that is the very package
you are trying to build.

Portage won't unpack a tar.xz archive with it's default tools
unless you inherit xz-utils.
But you can't write an ebuild that inherits itself.

I think in a RPM development environment, it's not as much of a sin
to presume you're bootstrapping most things.