[net-libs/xulrunner] Add EL SRPM-based ebuild for xulrunner.
Add Enterprise Linux SRPM-based Ebuild for net-libs/xulrunner-10.0.4.

The SRPM was sourced from Scientific Linux 6.2.
nss-3.13.1-7.el6_2.src.rpm

NOTES:
This ebuild is lightly tested.

It needs to be tested in conjunction with Enterprise Linux verions
of Firefox, nss, and nspr, as well as a few of the other packages
that would depend on xulrunner in the LTS6 environment.

Also, vendor default preference customizations are commented out.
They can be added back the next round of testing.