[net-mail/mailbase] Add EL SRPM-based ebuild for net-mail/mailbase.
Add Enterprise Linux SRPM-based Ebuild for net-mail/mailbase-1-r1.

The SRPM was sourced from Scientific Linux 6.2.

The 'aliases' file is a member of the setup package in EL SRPM:
setup-2.8.14-13.el6.src.rpm

The 'mailcap' file is a member of the mailcap package in EP SRPM:
mailcap-2.1.31-2.el6

The remaining files are managed in the ebuild 'FILESDIR' directory.