[sys-devel/libtool] Add EL SRPM-based ebuild for sys-devel/libtool.
Add Enterprise Linux SRPM-based Ebuild for sys-devel/libtool-2.2.6b-r11,

The SRPM was sourced from Scientific Linux 6.2.
libtool-2.2.6-15.5.el6.src.rpm

Maintainer Notes:
The SLOT for this package as it existed in the Gentoo CVS portage Attic
was "1.5".  I'm not sure of the original justification for slotting
this package as "1.5", and then adding blockers for co-existing with
other libtool-2* builds.

The slotting has been updated to SLOT="2" since this is consistent
with the slotting requirements for this version of libtool as
implemented in the context of the Enterprise Linux SRPM ebuilds.

Signed-off-by: Mitch Harder <mitch.harder@sabayonlinux.org>