[net-nds/openldap] Add EL SRPM-based ebuild for net-nds/openldap.
Add Enterprise Linux SRPM-based Ebuild for net-nds/openldap-2.4.23-r11

The SRPM was sourced from Scientific Linux 6.3.
openldap-2.4.23-26.el6_3.2.src.rpm

TO DO:
The implementation of smbk5pwd capabilities is different in the
Gentoo version as opposed to the Enterprise Linux version.

I have tried to keep the Gentoo implementation of smbk5pwd, but
it should be tested.

Signed-off-by: Mitch Harder <mitch.harder@sabayonlinux.org>